Final code.
The majority of the code was an attempt to use interrupts to do the job of detecting a switch press and pulsing a signal on another pin.  Ultimately I think the hardware I'm using is faulty because there was a 4-second delay between input captures, even when I stripped the code to the bare minimum.  Ultimately I used polling to do the trick.